refactor: split pulse.ts into 4 focused modules
- pulse.ts (176 lines): page layout, tile chrome (header/footer/wrap),
  auto-palette resolution, renderPulsePage
- pulse-renderers.ts (197 lines): 8 template renderers (metric,
  text-headline, table, status, time-series, image, text-image, media)
  + renderTile dispatcher
- pulse-helpers.ts (58 lines): esc, stringify, renderMarkdown,
  looksLikeJson, prettyJson
- pulse-templates.ts (225 lines): unchanged, template registry +
  normalizeSignal + extractMappedValues

Renderers receive a TileWrapFn callback to avoid circular imports.
